refactor: moved pages folder
