+++ src/frontend/fields/widget-fields.ts
@@ -8,7 +8,7 @@ import PanelField from './panel-field.vue';
 import ListField from './list-field.vue';
 import ObjectField from './object-field.vue';
 import ScriptureField from './scripture-field.vue';
-import BooleanField from './boolen-field.vue';
+import BooleanField from './boolean-field.vue';
 import SelectField from './select-field.vue';
 import NumberField from './number-field.vue';
